refactor: 리팩토링, 불필요한 인터페이스 삭제, 부적절한 infra 패키지 하위 구현체를 도메인 서비스로 올
