+++ app/src/main/java/com/joonhyeok/app/queue/domain/QueueFixedTotalPolicy.java
@@ -1,8 +1,5 @@
-package com.joonhyeok.app.queue.infra;
+package com.joonhyeok.app.queue.domain;
 
-import com.joonhyeok.app.queue.domain.Queue;
-import com.joonhyeok.app.queue.domain.QueuePolicy;
-import com.joonhyeok.app.queue.domain.QueueRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Limit;
 import org.springframework.stereotype.Service;
@@ -17,13 +14,12 @@ import static java.lang.Math.max;
 
 @Service
 @RequiredArgsConstructor
-public class QueueFixedTotalPolicy implements QueuePolicy {
+public class QueueFixedTotalPolicy  {
     private final Long FIXED_TOTAL = 5L;
     private final Long ACTIVE_WILL_EXPIRE_IN_MINUTES = 10L;
     private final QueueRepository queueRepository;
 
     @Transactional
-    @Override
     public void activate() {
         Long activateCount = queueRepository.countByStatus(ACTIVE);
         int limit = (int) max(0L, FIXED_TOTAL - activateCount);
@@ -32,7 +28,6 @@ public class QueueFixedTotalPolicy implements QueuePolicy {
     }
 
     @Transactional
-    @Override
     public void expire() {
         List<Queue> candidates = queueRepository.findByStatusAndExpireAtAfter(ACTIVE, LocalDateTime.now());
         candidates.forEach(Queue::expire);
