--- app/src/main/java/com/joonhyeok/app/queue/domain/QueuePolicy.java
@@ -1,6 +1,0 @@
-package com.joonhyeok.app.queue.domain;
-
-public interface QueuePolicy {
-    public void activate();
-    public void expire();
-}
