--- app/src/main/java/com/joonhyeok/app/reservation/domain/ReservationTimeOutPolicy.java
@@ -1,5 +1,0 @@
-package com.joonhyeok.app.reservation.domain;
-
-public interface ReservationTimeOutPolicy {
-    void invalidate();
-}
