+++ app/src/main/java/com/joonhyeok/app/user/application/UserPayService.java
@@ -8,11 +8,7 @@ import com.joonhyeok.app.reservation.domain.Reservation;
 import com.joonhyeok.app.reservation.domain.ReservationRepository;
 import com.joonhyeok.app.user.application.dto.user.UserPayCommand;
 import com.joonhyeok.app.user.application.dto.user.UserPayResult;
-import com.joonhyeok.app.user.domain.user.PayEvent;
-import com.joonhyeok.app.user.domain.user.PayEventPublisher;
-import com.joonhyeok.app.user.domain.user.PayValidator;
-import com.joonhyeok.app.user.domain.user.User;
-import com.joonhyeok.app.user.domain.user.UserRepository;
+import com.joonhyeok.app.user.domain.user.*;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
