+++ app/src/test/java/com/joonhyeok/app/user/UserPayServiceIntegrateTest.java
@@ -6,19 +6,15 @@ import com.joonhyeok.app.concert.domain.SeatStatus;
 import com.joonhyeok.app.reservation.domain.Reservation;
 import com.joonhyeok.app.reservation.domain.ReservationRepository;
 import com.joonhyeok.app.reservation.domain.ReservationStatus;
+import com.joonhyeok.app.user.application.OutboxService;
 import com.joonhyeok.app.user.application.UserPayService;
+import com.joonhyeok.app.user.application.dto.outbox.OutboxFindCommand;
 import com.joonhyeok.app.user.application.dto.user.UserPayCommand;
 import com.joonhyeok.app.user.application.dto.user.UserPayResult;
-import com.joonhyeok.app.user.infra.domain.outbox.PayOutboxKafkaListener;
-import com.joonhyeok.app.user.domain.user.PayEvent;
-import com.joonhyeok.app.user.domain.user.PayEventListener;
-import com.joonhyeok.app.user.domain.user.Account;
-import com.joonhyeok.app.user.domain.user.User;
-import com.joonhyeok.app.user.domain.user.UserRepository;
 import com.joonhyeok.app.user.domain.outbox.Outbox;
-import com.joonhyeok.app.user.application.dto.outbox.OutboxFindCommand;
-import com.joonhyeok.app.user.application.OutboxService;
 import com.joonhyeok.app.user.domain.outbox.OutboxStatus;
+import com.joonhyeok.app.user.domain.user.*;
+import com.joonhyeok.app.user.infra.domain.outbox.PayOutboxKafkaListener;
 import io.zonky.test.db.AutoConfigureEmbeddedDatabase;
 import jakarta.persistence.EntityNotFoundException;
 import org.junit.jupiter.api.DisplayName;
